requirements.txt
astroid==2.15.6
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
dill==0.3.6
flake8==6.0.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
pip==23.1.2
platformdirs==3.8.1
pluggy==1.2.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.17.4
pytest==7.4.0
requests==2.31.0
rsa==4.9
setuptools==68.0.0
six==1.16.0
soupsieve==2.4.1
tomlkit==0.11.8
urllib3==1.26.16
wheel==0.40.0
wrapt==1.15.0
